Memory usage optimisation for trie-building.
When adding strings to the trie, we add the "normal" string followed by an uppercase copy.

This optimisation uses the knowledge that the trie-building algorithm only looks at the
keys character by character (via operator[]). By altering the output of this function
to uppercase characters on-the-fly we can avoid adding a second copy of the string value
to the Keyset.

Note; this only works when the uppercasing is possible with toupper, i.e. plain ASCII.

All the data classes (marisa::Key, Entry, Key, ReverseKey) are modified to contain a boolean
upper field, which is only true for this particular case. To "make space" for this field
the length has been dropped from 32-bit to 16-bit field. This is sufficient for us since
we know the strings are max 4KB in size (which itself is generous) but it's definitely a
restriction.

An optional alternative implementation could store this value as a single bit leaving
31 bits remaining for the length, at the expense of some bit-twiddling in various places.